패키지 별 묶기, 불필요 컨트롤러 제거
