--- src/main/java/com/ll/spring_additional/CommonUtil.java
@@ -1,16 +1,0 @@
-package com.ll.spring_additional;
-
-import org.commonmark.node.Node;
-import org.commonmark.parser.Parser;
-import org.commonmark.renderer.html.HtmlRenderer;
-import org.springframework.stereotype.Component;
-
-@Component
-public class CommonUtil {
-	public String markdown(String markdown) {
-		Parser parser = Parser.builder().build();
-		Node document = parser.parse(markdown);
-		HtmlRenderer renderer = HtmlRenderer.builder().build();
-		return renderer.render(document);
-	}
-}
