--- src/main/java/com/ll/spring_additional/HelloController.java
@@ -1,14 +1,0 @@
-package com.ll.spring_additional;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class HelloController {
-	@GetMapping("/hello")
-	@ResponseBody
-	public String hello() {
-		return "hello world";
-	}
-}
