--- src/main/java/com/ll/spring_additional/answer/AnswerForm.java
@@ -1,12 +1,0 @@
-package com.ll.spring_additional.answer;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AnswerForm {
-	@NotEmpty(message = "내용은 필수항목입니다.")
-	private String content;
-}
