--- src/main/java/com/ll/spring_additional/answer/AnswerRepository.java
@@ -1,7 +1,0 @@
-package com.ll.spring_additional.answer;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AnswerRepository extends JpaRepository<Answer, Integer> {
-
-}
