+++ src/main/java/com/ll/spring_additional/boundedContext/answer/form/AnswerForm.java
@@ -1,0 +1,12 @@
+package com.ll.spring_additional.boundedContext.answer.form;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AnswerForm {
+	@NotEmpty(message = "내용은 필수항목입니다.")
+	private String content;
+}
