+++ src/main/java/com/ll/spring_additional/boundedContext/answer/repository/AnswerRepository.java
@@ -1,0 +1,9 @@
+package com.ll.spring_additional.boundedContext.answer.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.ll.spring_additional.boundedContext.answer.entity.Answer;
+
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {
+
+}
