+++ src/main/java/com/ll/spring_additional/boundedContext/home/controller/HomeController.java
@@ -1,0 +1,13 @@
+package com.ll.spring_additional.boundedContext.home.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+public class HomeController {
+	@GetMapping("/")
+	public String root() {
+		return "redirect:/question/list";
+	}
+}
