+++ src/main/java/com/ll/spring_additional/boundedContext/user/UserRole.java
@@ -1,0 +1,15 @@
+package com.ll.spring_additional.boundedContext.user;
+
+import lombok.Getter;
+
+@Getter
+public enum UserRole {
+	ADMIN("ROLE_ADMIN"),
+	USER("ROLE_USER");
+
+	UserRole(String value) {
+		this.value = value;
+	}
+
+	private String value;
+}
