+++ src/main/java/com/ll/spring_additional/boundedContext/user/controller/UserController.java
@@ -1,0 +1,59 @@
+package com.ll.spring_additional.boundedContext.user.controller;
+import jakarta.validation.Valid;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import com.ll.spring_additional.boundedContext.user.Form.UserCreateForm;
+import com.ll.spring_additional.boundedContext.user.service.UserService;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Controller
+@RequestMapping("/user")
+public class UserController {
+
+	private final UserService userService;
+
+	@GetMapping("/login")
+	public String login() {
+		return "user/login_form";
+	}
+
+	@GetMapping("/signup")
+	public String signup(UserCreateForm userCreateForm) {
+		return "user/signup_form";
+	}
+
+	@PostMapping("/signup")
+	public String signup(@Valid UserCreateForm userCreateForm, BindingResult bindingResult) {
+		if (bindingResult.hasErrors()) {
+			return "user/signup_form";
+		}
+
+		if (!userCreateForm.getPassword1().equals(userCreateForm.getPassword2())) {
+			bindingResult.rejectValue("password2", "passwordInCorrect",
+				"2개의 패스워드가 일치하지 않습니다.");
+			return "user/signup_form";
+		}
+
+		try {
+			userService.create(userCreateForm.getUsername(),
+				userCreateForm.getEmail(), userCreateForm.getPassword1());
+		}catch(DataIntegrityViolationException e) {
+			e.printStackTrace();
+			bindingResult.reject("signupFailed", "이미 등록된 사용자입니다.");
+			return "user/signup_form";
+		}catch(Exception e) {
+			e.printStackTrace();
+			bindingResult.reject("signupFailed", e.getMessage());
+			return "user/signup_form";
+		}
+		return "redirect:/";
+	}
+}
