+++ src/main/java/com/ll/spring_additional/boundedContext/user/entity/SiteUser.java
@@ -1,0 +1,27 @@
+package com.ll.spring_additional.boundedContext.user.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Entity
+public class SiteUser {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@Column(unique = true)
+	private String username;
+
+	private String password;
+
+	@Column(unique = true)
+	private String email;
+}
