--- src/main/java/com/ll/spring_additional/question/QuestionService.java
@@ -1,68 +1,0 @@
-package com.ll.spring_additional.question;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.ll.spring_additional.DataNotFoundException;
-import com.ll.spring_additional.user.SiteUser;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class QuestionService {
-	private final QuestionRepository questionRepository;
-
-	public Page<Question> getList(int page, String kw) {
-		List<Sort.Order> sorts = new ArrayList<>();
-		sorts.add(Sort.Order.desc("createDate"));
-		Pageable pageable = PageRequest.of(page, 10, Sort.by(sorts)); //페이지 번호, 개수
-		return questionRepository.findAllByKeyword(kw, pageable);
-	}
-
-	public Question getQuestion(Integer id) {
-		Optional<Question> question = questionRepository.findById(id);
-		if (question.isPresent()) {
-			return question.get();
-		} else {
-			throw new DataNotFoundException("question not found");
-		}
-	}
-
-	@Transactional
-	public void create(String subject, String content, SiteUser user) {
-		Question q = new Question();
-		q.setSubject(subject);
-		q.setContent(content);
-		q.setAuthor(user);
-		questionRepository.save(q);
-	}
-
-	@Transactional
-	public void modify(Question question, String subject, String content) {
-		question.setSubject(subject);
-		question.setContent(content);
-		questionRepository.save(question);
-	}
-
-	@Transactional
-	public void delete(Question question) {
-		questionRepository.delete(question);
-	}
-
-	@Transactional
-	public void vote(Question question, SiteUser siteUser) {
-		question.getVoters().add(siteUser);
-		questionRepository.save(question);
-	}
-}
