--- src/main/java/com/ll/spring_additional/user/UserRole.java
@@ -1,15 +1,0 @@
-package com.ll.spring_additional.user;
-
-import lombok.Getter;
-
-@Getter
-public enum UserRole {
-	ADMIN("ROLE_ADMIN"),
-	USER("ROLE_USER");
-
-	UserRole(String value) {
-		this.value = value;
-	}
-
-	private String value;
-}
