--- src/main/java/com/ll/spring_additional/user/UserSecurityService.java
@@ -1,39 +1,0 @@
-package com.ll.spring_additional.user;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-public class UserSecurityService implements UserDetailsService {
-
-	private final UserRepository userRepository;
-	// 사용자명으로 비밀번호를 조회하여 리턴하는 메서드
-
-	@Override
-	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		Optional<SiteUser> _siteUser = this.userRepository.findByUsername(username);
-		if (_siteUser.isEmpty()) {
-			throw new UsernameNotFoundException("사용자를 찾을수 없습니다.");
-		}
-		SiteUser siteUser = _siteUser.get();
-		List<GrantedAuthority> authorities = new ArrayList<>();
-		if ("admin".equals(username)) {
-			authorities.add(new SimpleGrantedAuthority(UserRole.ADMIN.getValue()));
-		} else {
-			authorities.add(new SimpleGrantedAuthority(UserRole.USER.getValue()));
-		}
-		return new User(siteUser.getUsername(), siteUser.getPassword(), authorities);
-	}
-}
