--- src/main/java/com/ll/spring_additional/user/UserService.java
@@ -1,41 +1,0 @@
-package com.ll.spring_additional.user;
-
-import java.util.Optional;
-
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import com.ll.spring_additional.DataNotFoundException;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Service
-@Transactional(readOnly = true)
-public class UserService {
-
-	private final UserRepository userRepository;
-
-	private final PasswordEncoder passwordEncoder;
-
-	@Transactional
-	public SiteUser create(String username, String email, String password) {
-		SiteUser user = new SiteUser();
-		user.setUsername(username);
-		user.setEmail(email);
-		user.setPassword(passwordEncoder.encode(password));
-		userRepository.save(user);
-		return user;
-	}
-
-	public SiteUser getUser(String username) {
-		Optional<SiteUser> siteUser = userRepository.findByusername(username);
-		if (siteUser.isPresent()) {
-			return siteUser.get();
-		} else {
-			throw new DataNotFoundException("siteuser not found");
-		}
-	}
-}
