+++ src/main/resources/templates/answer/answer_form.html
@@ -1,0 +1,13 @@
+<html layout:decorate="~{common/layout}">
+<div layout:fragment="content" class="container">
+    <h5 class="my-3 border-bottom pb-2">답변 수정</h5>
+    <form th:object="${answerForm}" th:action method="post">
+        <div th:replace="~{common/form_errors :: formErrorsFragment}"></div>
+        <div class="mb-3">
+            <label for="content" class="form-label">내용</label>
+            <textarea th:field="*{content}" class="form-control" rows="10"></textarea>
+        </div>
+        <input type="submit" value="저장하기" class="btn btn-primary my-2">
+    </form>
+</div>
+</html>
