+++ src/main/resources/templates/common/navbar.html
