--- src/main/resources/templates/question_form.html
@@ -1,17 +1,0 @@
-<html layout:decorate="~{layout}">
-<div layout:fragment="content" class="container">
-  <h5 class="my-3 border-bottom pb-2">질문등록</h5>
-  <form th:action th:object="${questionForm}" method="post">
-    <div th:replace="~{form_errors :: formErrorsFragment}"></div>
-    <div class="mb-3">
-      <label for="subject" class="form-label">제목</label>
-      <input type="text" th:field="*{subject}" class="form-control">
-    </div>
-    <div class="mb-3">
-      <label for="content" class="form-label">내용</label>
-      <textarea th:field="*{content}" class="form-control" rows="10"></textarea>
-    </div>
-    <input type="submit" value="저장하기" class="btn btn-primary my-2">
-  </form>
-</div>
-</html>
