--- src/main/resources/templates/question_list.html
@@ -1,75 +1,0 @@
-<html layout:decorate="~{layout}">
-<div layout:fragment="content" class="container my-3">
-    <div class="row my-3">
-        <div class="col-6">
-            <a th:href="@{/question/create}" class="btn btn-primary">질문 등록하기</a>
-        </div>
-        <div class="col-6">
-            <form>
-            <div class="input-group">
-                <input type="text" name="kw" class="form-control" th:value="${param.kw}">
-                <button class="btn btn-outline-secondary" type="button" id="btn_search">찾기</button>
-            </div>
-            </form>
-        </div>
-    </div>
-    <table class="table">
-        <thead class="table-dark">
-        <tr class="text-center">
-            <th>번호</th>
-            <th style="width:50%">제목</th>
-            <th>글쓴이</th>
-            <th>작성일시</th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr class="text-center" th:each="question, loop : ${paging}">
-            <td th:text="${paging.getTotalElements - (paging.number * paging.size) - loop.index}"></td>
-            <td class="text-start">
-                <a th:href="@{|/question/detail/${question.id}|}" th:text="${question.subject}"></a>
-                <span class="text-danger small ms-2"
-                      th:if="${#lists.size(question.answerList) > 0}"
-                      th:text="${#lists.size(question.answerList)}">
-                </span>
-            <td><span th:if="${question.author != null}" th:text="${question.author.username}"></span></td>
-            <td th:text="${#temporals.format(question.createDate, 'yyyy-MM-dd HH:mm')}"></td>
-        </tr>
-        </tbody>
-    </table>
-    <!-- 페이징처리 시작 -->
-    <div th:if="${!paging.isEmpty()}" th:with="queryStrBase = '?kw=' + ${param.kw != null ? param.kw : ''}">
-        <ul class="pagination justify-content-center">
-            <li class="page-item" th:classappend="${paging.number==0} ? 'disabled'">
-                <a class="page-link"
-                   th:href="@{|${queryStrBase}&page=0|}">
-                    <span>처음</span>
-                </a>
-            </li>
-            <li class="page-item" th:classappend="${!paging.hasPrevious} ? 'disabled'">
-                <a class="page-link"
-                   th:href="@{|${queryStrBase}&page=${paging.number-1}|}">
-                    <span>이전</span>
-                </a>
-            </li>
-            <li th:each="page: ${#numbers.sequence(0, paging.totalPages-1)}"
-                th:if="${page >= paging.number-5 and page <= paging.number+5}"
-                th:classappend="${page == paging.number} ? 'active'"
-                class="page-item">
-                <a th:text="${page}" class="page-link" th:href="@{|${queryStrBase}&page=${page}|}"></a>
-            </li>
-            <li class="page-item" th:classappend="${!paging.hasNext} ? 'disabled'">
-                <a class="page-link" th:href="@{|${queryStrBase}&page=${paging.number+1}|}">
-                    <span>다음</span>
-                </a>
-            </li>
-            <li class="page-item" th:classappend="${paging.number==(paging.totalPages-1)} ? 'disabled'">
-                <a class="page-link"
-                   th:href="@{|${queryStrBase}&page=${paging.totalPages-1}|}">
-                    <span>끝</span>
-                </a>
-            </li>
-        </ul>
-    </div>
-    <!-- 페이징처리 끝 -->
-</div>
-</html>
