+++ src/main/resources/templates/user/login_form.html
@@ -1,0 +1,21 @@
+<html layout:decorate="~{common/layout}">
+<div layout:fragment="content" class="container my-3">
+  <form th:action="@{/user/login}" method="post">
+    <!-- 로그인 실패 시 파라미터로 error 전달, 스프링 시큐리티 규칙 -->
+    <div th:if="${param.error}">
+      <div class="alert alert-danger">
+        사용자ID 또는 비밀번호를 확인해 주세요.
+      </div>
+    </div>
+    <div class="mb-3">
+      <label for="username" class="form-label">사용자ID</label>
+      <input type="text" name="username" id="username" class="form-control">
+    </div>
+    <div class="mb-3">
+      <label for="password" class="form-label">비밀번호</label>
+      <input type="password" name="password" id="password" class="form-control">
+    </div>
+    <button type="submit" class="btn btn-primary">로그인</button>
+  </form>
+</div>
+</html>
