+++ src/main/resources/templates/user/signup_form.html
@@ -1,0 +1,29 @@
+<html layout:decorate="~{common/layout}">
+<div layout:fragment="content" class="container my-3">
+  <div class="my-3 border-bottom">
+    <div>
+      <h4>회원가입</h4>
+    </div>
+  </div>
+  <form th:action="@{/user/signup}" th:object="${userCreateForm}" method="post">
+    <div th:replace="~{common/form_errors :: formErrorsFragment}"></div>
+    <div class="mb-3">
+      <label for="username" class="form-label">사용자ID</label>
+      <input type="text" th:field="*{username}" class="form-control">
+    </div>
+    <div class="mb-3">
+      <label for="password1" class="form-label">비밀번호</label>
+      <input type="password" th:field="*{password1}" class="form-control">
+    </div>
+    <div class="mb-3">
+      <label for="password2" class="form-label">비밀번호 확인</label>
+      <input type="password" th:field="*{password2}" class="form-control">
+    </div>
+    <div class="mb-3">
+      <label for="email" class="form-label">이메일</label>
+      <input type="email" th:field="*{email}" class="form-control">
+    </div>
+    <button type="submit" class="btn btn-primary">회원가입</button>
+  </form>
+</div>
+</html>
